requirements.txt
Django>=4.2
psycopg2-binary
dj-database-url
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
whitenoise
python-dotenv